[FIX] fix fork and lfork

+++ src/instructions/fork.c
@@ -33,7 +33,7 @@ int execute_fork(corewar_t *cw, champions_t *c, int ins, int *args)
     cw->status_champ = realloc(cw->status_champ, sizeof(bool) * (cw->\
     nbr_champions + 1));
     cw->status_champ[cw->nbr_champions - 1] = true;
-    cw->live_call = realloc(cw->live_call, sizeof(int) * (cw->\
+    cw->live_call = realloc(cw->live_call, sizeof(size_t) * (cw->\
     nbr_champions + 2));
     cw->live_call[cw->nbr_champions - 1] = 0;
     cw->champions[cw->nbr_champions] = malloc(sizeof(champions_t));
